Added e2e test.

+++ pkg/handlers/nfs_handler.go
@@ -116,7 +116,7 @@ func (h *nfsHandler) OnAdd(ns string, vc kvcv1.VolumeConfig, controllerRef metav
 
 func (h *nfsHandler) OnDelete(ns string, vc kvcv1.VolumeConfig, controllerRef metav1.OwnerReference) {
 	for _, client := range h.k8sResourceClients {
-		if client.Plural() == "nodes" {
+		if client.Plural() == "nodes" || client.Plural() == "pods" {
 			continue
 		}
 
@@ -127,6 +127,10 @@ func (h *nfsHandler) OnDelete(ns string, vc kvcv1.VolumeConfig, controllerRef me
 
 		for _, resource := range resourceList {
 			resControllerRef := metav1.GetControllerOf(resource)
+			if resControllerRef == nil {
+				continue
+			}
+
 			if resControllerRef.UID == controllerRef.UID {
 				client.Delete(ns, resource.GetName())
 			}
